+++ public/index.html
@@ -48,11 +48,17 @@
     <link rel="stylesheet" href="./css/main.css" />
     <link rel="stylesheet" type="text/css" href="css/main.css" />
     <script defer type="module">
-      import app from './js/app.js'; //This line imports the app
-      app.init('Chingu Task Tracker'); // This line initializes the app
+      import app from "./js/app.js"; //This line imports the app
+      app.init("Chingu Task Tracker"); // This line initializes the app
       //Value of 'Chingu Task Tracker' passes title into app, default is 'Daily Checklist'
       console.log(app); //Provides user with app level information inside of the console
       //Should remove console.log before going live
+      //console.log(app.controller.loadData());
+      //const task = app.tasks[0];
+      //const {group, category} = task;
+      //  console.log(app.view.asideViewController.renderAside());
+      // console.log(app.view.asideViewController.renderAsideGroups());
+      // console.log(app.view.asideViewController.renderAsideCategory(category,group));
     </script>
   </head>
   <body></body>
