--- public/js/views/content/renderContent.js
@@ -1,29 +1,0 @@
-import app from "../../app.js";
-import renderContentGroups from "./renderContentGroups.js";
-import renderContentTasks from "./renderContentTasks.js";
-
-const renderContent = () => {
-  app.view.createElement(
-    "content",
-    "",
-    document.getElementById("contentAnchor"),
-    "content",
-    "content"
-  );
-  renderContentGroups();
-  renderContentTasks();
-};
-
-export const updateContent = () => {
-  document.getElementById("content").remove();
-  renderContent();
-};
-
-export const renderFilteredTasks = (tasks) => {
-  let temp = app.tasks;
-  app.tasks = tasks;
-  updateContent(tasks);
-  app.tasks = temp;
-};
-
-export default renderContent;
