+++ public/js/views/view.index.js
@@ -1,10 +1,7 @@
+import AppViewController from "./AppView.js";
 import AsideViewController from "./AsideView.js";
+import ContentViewController from "./ContentView.js";
 import TaskViewController from "./TaskView.js";
-import renderNavBar from "./renderNavbar.js";
-import renderContent, { updateContent } from "./content/renderContent.js";
-import { removePopup } from "./modals/view-task.js";
-import renderFooter from "./renderFooter.js";
-import { renderModalButton } from "./modals/add-edit-task.js";
 
 let viewIndex = 1;
 export default class View {
@@ -26,36 +23,15 @@ export default class View {
     anchor.append(container); //Where we are appending the container too
     this.container = container; //Stores container on obj so we can access later
   }
-  asideViewController = new AsideViewController();
-  taskViewController = new TaskViewController();
-  init(title) {
-    this.asideViewController.init(title);
-    renderNavBar();
-    renderContent();
-    renderFooter();
-    renderModalButton();
-  }
   // createElement(element (what type of element is is ie div or footer): any, content (what is the inner html): any, anchor (what are we apending it to, where we are putting the element, it goes inside whatever we put here): any, id (optional, sets the id): any, classList (optional, sets the classlist): any): void
   //Function to display the data into HTML:
   createElement(element, content, anchor, id, classList) {
     return new View(element, content, anchor, id, classList);
   }
-  updateView(filteredTasks) {
-    let temp = app.tasks; //Defines a temporary variable to store all tasks on
-    if (filteredTasks) app.tasks = filteredTasks; //Checking if we have a filtered array from user search, if so assign app.tasks to that array
-    removePopup(); //Removes the modal popup
-    this.asideViewController.updateAsideGroups();
-    updateContent();
-    app.tasks = temp;
-    if (app.tasks && app.tasks.length > 0) {
-      if (document.getElementById("search").classList.includes("active"))
-        return;
-      else document.getElementById("search").value = "";
-    }
-  }
-  toggleCategory(group, category) {
-    //console.log(group,category, app.controller.returnCategoryByGroup(group).map(task => task.toggleComplete()));
-  }
+  appViewController = new AppViewController();
+  asideViewController = new AsideViewController();
+  contentViewController = new ContentViewController();
+  taskViewController = new TaskViewController();
 }
 
 // Hamburger Menu Display on Mobile < Emmetts code for hamburger menu
